OUA: OUA Customisations and plugin additions to Moodle.
As part of the handover to eCreators, only the changes are able to be
submitted.  The history is not available due to other constraints about
code that was in the codebase.  These are all applied in a single commit
to a base Moodle installation at the point of last upgrade.

Some historical conflicts may re-appear when merging future patches.  This
is unlikely as Modoel 3.0 is not in security fix only support.